+2006-01-14  Junichi Uekawa  <dancer@debian.org> + +	* tests/016_memleakcheck.sh: use mcheck to see if memory leaks +	happen. + +	* tests/015_test_ilistdelete.sh: disabled the check since it's +	difficult to formulate the test yet. + +	* tests/002_test_open.c (main): fix thinko. It was never tested, +	really. + +2006-01-13  Junichi Uekawa  <dancer@debian.org> + +	* tests/015_test_ilistdelete.c (main): a preliminary +	implementation, this is wrong. +	I should also test pthread and doing something racy. + +	* cowdancer.c (initialize_functions): logic error, if someone else +	is initializing and it has failed, we'd probably need to retry, +	not hang. + +	* tests/015_test_ilistdelete.sh: test ilist delete code path. +

+++ tests/014_ilistdump_test.c
@@ -24,7 +24,7 @@
 int main(int argc, char** argv)
 {
   struct ilist_struct s;
-  fprintf (stderr, "ilist_struct size on this architecture: %i\n", sizeof(struct ilist_struct));
+  fprintf (stderr, "ilist_struct size on this architecture: %i\n", (int)sizeof(struct ilist_struct));
   while(fread(&s, sizeof(struct ilist_struct), 1, stdin))
     {
       printf ("%li %li\n", (long)s.dev, (long)s.inode);
